requirements.txt
dspy-ai>=2.4.0 
langgraph>=0.1.0 
langchain-core>=0.2.0 
pydantic>=2.0.0 
click>=8.1.7 
rich>=13.7.0 
numpy>=1.26.0 
pandas>=2.2.0 
scikit-learn>=1.3.0 
rank-bm25>=0.2.2  # optional 